% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\encoding{UTF-8}
\name{bde_catalog_load}
\alias{bde_catalog_load}
\title{Load BdE catalog metadata}
\source{
\href{https://www.bde.es/webbe/en/estadisticas/recursos/descargas-completas.html}{time series bulk data download}.
}
\usage{
bde_catalog_load(
  catalog = c("ALL", "BE", "SI", "TC", "TI", "PB"),
  parse_dates = TRUE,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{catalog}{A single catalog identifier to load, or \code{"ALL"} to load every
catalog. See \strong{Details}.}

\item{parse_dates}{Logical. If \code{TRUE}, date columns are parsed with
\code{\link[=bde_parse_dates]{bde_parse_dates()}}.}

\item{cache_dir}{A path to a cache directory. The directory can also be set
with options using \code{options(bde_cache_dir = "path/to/dir")}.}

\item{update_cache}{Logical. If \code{TRUE}, the requested file is refreshed in
\code{cache_dir}.}

\item{verbose}{Logical. If \code{TRUE}, display information useful for debugging.}
}
\value{
A \link[tibble:tbl_df]{tibble} with the requested catalog metadata.
}
\description{
Load BdE time series catalog metadata.
}
\details{
Accepted values for \code{catalog} are:\tabular{llll}{
   \strong{CODE} \tab \strong{PUBLICATION} \tab \strong{UPDATE FREQUENCY} \tab \strong{FREQUENCY} \cr
   \code{"BE"} \tab Statistical Bulletin \tab Daily \tab Monthly \cr
   \code{"SI"} \tab Summary Indicators \tab Daily \tab Daily \cr
   \code{"TC"} \tab Exchange Rates \tab Daily \tab Daily \cr
   \code{"TI"} \tab Interest Rates \tab Daily \tab Daily \cr
   \code{"PB"} \tab Bank Lending Survey \tab Quarterly \tab Quarterly \cr
}


Use \code{"ALL"} as a shorthand for loading all catalogs at once.

If the requested catalog is not cached, this function calls
\code{\link[=bde_catalog_update]{bde_catalog_update()}}.
}
\examples{
\dontshow{if (bde_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
bde_catalog_load("TI", verbose = TRUE)
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other catalog:
\code{\link[=bde_catalog_search]{bde_catalog_search()}},
\code{\link[=bde_catalog_update]{bde_catalog_update()}}
}
\concept{catalog}
