% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\encoding{UTF-8}
\name{bde_catalog_update}
\alias{bde_catalog_update}
\title{Update BdE catalog files}
\source{
\href{https://www.bde.es/webbe/en/estadisticas/recursos/descargas-completas.html}{time series bulk data download}.
}
\usage{
bde_catalog_update(
  catalog = c("ALL", "BE", "SI", "TC", "TI", "PB"),
  cache_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{catalog}{A single catalog identifier to update, or \code{"ALL"} to update
every catalog. See \strong{Details}.}

\item{cache_dir}{A path to a cache directory. The directory can also be set
with options using \code{options(bde_cache_dir = "path/to/dir")}.}

\item{verbose}{Logical. If \code{TRUE}, display information useful for debugging.}
}
\value{
An invisible list of download results.
}
\description{
Update BdE time series catalog files.
}
\details{
Accepted values for \code{catalog} are:\tabular{llll}{
   \strong{CODE} \tab \strong{PUBLICATION} \tab \strong{UPDATE FREQUENCY} \tab \strong{FREQUENCY} \cr
   \code{"BE"} \tab Statistical Bulletin \tab Daily \tab Monthly \cr
   \code{"SI"} \tab Summary Indicators \tab Daily \tab Daily \cr
   \code{"TC"} \tab Exchange Rates \tab Daily \tab Daily \cr
   \code{"TI"} \tab Interest Rates \tab Daily \tab Daily \cr
   \code{"PB"} \tab Bank Lending Survey \tab Quarterly \tab Quarterly \cr
}


Use \code{"ALL"} as a shorthand for updating all catalogs at once.
}
\examples{
\dontshow{if (bde_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
bde_catalog_update("TI", verbose = TRUE)
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other catalog:
\code{\link[=bde_catalog_load]{bde_catalog_load()}},
\code{\link[=bde_catalog_search]{bde_catalog_search()}}
}
\concept{catalog}
