% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{bde_parse_dates}
\alias{bde_parse_dates}
\title{Parse date strings}
\usage{
bde_parse_dates(dates_to_parse)
}
\arguments{
\item{dates_to_parse}{A character vector of date strings to parse.}
}
\value{
A vector of \code{\link[=as.Date]{Date}} values.
}
\description{
Parses date strings with \code{\link[=as.Date]{as.Date()}}. This function is tailored to the date
formats used in this package and may not parse other datasets. See
\strong{Examples} for supported formats.
}
\details{
\subsection{Date formats}{\tabular{lll}{
   \strong{FREQUENCY} \tab \strong{FORMAT} \tab \strong{EXAMPLES} \cr
   \strong{Daily / Business day} \tab \verb{DD MMMYYYY} \tab \verb{02 FEB2019} \cr
   \strong{Monthly} \tab \verb{MMM YYYY} \tab \verb{MAR 2020} \cr
   \strong{Quarterly} \tab \verb{MMM YYYY}, where \code{MMM} is the first or last month of the quarter, depending on the value of the \code{OBSERVED} variable. \tab For the first quarter of 2020: \verb{ENE 2020}, \verb{MAR 2020} \cr
   \strong{Half-yearly} \tab \verb{MMM YYYY}, where \code{MMM} is the first or last month of the half-year period, depending on the value of its \code{OBSERVED} variable. \tab For the first half of 2020: \verb{ENE 2020}, \verb{JUN 2020} \cr
   \strong{Annual} \tab \code{YYYY} \tab \code{2020} \cr
}


See \code{vignette("csv_manual", package = "tidyBdE")} for details.
}
}
\examples{
# Supported formats.
would_parse <- c(
  "02 FEB2019", "15 ABR 1890", "MAR 2020", "ENE2020",
  "2020", "12-1993", "01-02-2014", "01/02/1990"
)

parsed_ok <- bde_parse_dates(would_parse)

class(parsed_ok)

dplyr::tibble(raw = would_parse, parsed = parsed_ok)

# Unsupported formats.

wont_parse <- c("JAN2001", "2010-01-12", "01 APR 2017", "01/31/1990")

parsed_fail <- bde_parse_dates(wont_parse)

class(parsed_fail)

dplyr::tibble(raw = wont_parse, parsed = parsed_fail)
}
\seealso{
\itemize{
\item \code{\link[=bde_catalog_load]{bde_catalog_load()}} and \code{\link[=bde_series_load]{bde_series_load()}} use this parser.
\item \code{\link[=as.Date]{as.Date()}} provides base R date conversion.
}
}
\concept{utils}
