% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series.R
\encoding{UTF-8}
\name{bde_series_full_load}
\alias{bde_series_full_load}
\title{Load BdE full time series files}
\usage{
bde_series_full_load(
  series_csv,
  parse_dates = TRUE,
  parse_numeric = TRUE,
  cache_dir = NULL,
  update_cache = FALSE,
  verbose = FALSE,
  extract_metadata = FALSE
)
}
\arguments{
\item{series_csv}{CSV file of a series, as defined in the field
\verb{Nombre del archivo con los valores de la serie} of the corresponding
catalog. See \code{\link[=bde_catalog_load]{bde_catalog_load()}}.}

\item{parse_dates}{Logical. If \code{TRUE}, date columns are parsed with
\code{\link[=bde_parse_dates]{bde_parse_dates()}}.}

\item{parse_numeric}{Logical. If \code{TRUE}, the columns are parsed to
double (numeric) values. See \strong{Note}.}

\item{cache_dir}{A path to a cache directory. The directory can also be set
with options using \code{options(bde_cache_dir = "path/to/dir")}.}

\item{update_cache}{Logical. If \code{TRUE}, the requested file is refreshed in
\code{cache_dir}.}

\item{verbose}{Logical. If \code{TRUE}, display information useful for debugging.}

\item{extract_metadata}{Logical. If \code{TRUE}, the output is the metadata of the
requested series.}
}
\value{
A \link[tibble:tbl_df]{tibble} with a \code{Date} field and the aliases of the
series fields as described in the catalogs. See \code{\link[=bde_catalog_load]{bde_catalog_load()}}.
}
\description{
Load a full BdE time series file.
}
\details{
\subsection{About BdE file naming}{

The series name is a positional code showing the location of the table. For
example, table \strong{be_6_1} represents Table 1, Chapter 6 of the Statistical
Bulletin ("BE"). Although it is unique, it is subject to change, for
example when a new table is inserted before it.

For that reason, \code{\link[=bde_series_load]{bde_series_load()}} is more suitable for extracting
specific time series.
}
}
\note{
This function tries to coerce the columns to numbers. For some series, a
warning may be displayed if the parser fails. You can override the default
behavior with \code{parse_numeric = FALSE}.
}
\examples{
\dontshow{if (bde_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
# Show metadata.
bde_series_full_load("TI_1_1.csv", extract_metadata = TRUE)

# Load data.
bde_series_full_load("TI_1_1.csv")
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other series:
\code{\link[=bde_series_load]{bde_series_load()}}
}
\concept{series}
