% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic.R
\encoding{UTF-8}
\name{bde_themes_download}
\alias{bde_themes_download}
\title{Batch download of files included on the thematic classification}
\source{
\href{https://www.bde.es/webbde/en/estadis/infoest/descarga_series_temporales.html}{Time-series bulk data download}
}
\usage{
bde_themes_download(themes = "ALL", cache_dir = NULL, verbose = FALSE)
}
\arguments{
\item{themes}{A single value indicating the themes to be downloaded
or \code{"ALL"} as a shorthand. See \strong{Details}.}

\item{cache_dir}{A path to a cache directory. The directory can also be set
via options with \code{options(bde_cache_dir = "path/to/dir")}.}

\item{verbose}{Logical \code{TRUE} or \code{FALSE}, display information useful for
debugging.}
}
\value{
Invisible. This function is called for its side effects.
}
\description{
By the time, this function just batch-download all the series included
in the
\href{https://www.bde.es/webbde/en/estadis/infoest/descarga_series_temporales.html}{Theme-based classification}
provided by Bank of Spain. However, this is useful as the function
\code{\link[=bde_series_load_all]{bde_series_load_all()}} and \code{\link[=bde_series_load]{bde_series_load()}} would search also in the
downloaded files, providing a faster response to the end user.
}
\details{
Download several series based on the thematic classification of BdE.

Accepted values for \code{themes} are:\tabular{ll}{
   themes \tab Description \cr
   \strong{TE_TIPOS} \tab Interest rate statistics \cr
   \strong{TE_TIPOSCAM} \tab Exchange rate statistics \cr
   \strong{TE_CF} \tab Financial accounts statistics \cr
   \strong{TE_IFINA} \tab Financial corporations statistics \cr
   \strong{TE_DEU} \tab General government statistics \cr
   \strong{TE_CENBAL} \tab Non-Financial corporations statistics \cr
   \strong{TE_SECEXT} \tab External statistics \cr
   \strong{TE_MERCA} \tab Financial market statistics \cr
   \strong{TE_ESTECON} \tab General economic statistics \cr
}


Use \code{"ALL"} as a shorthand for updating all the catalogs at a glance.

If this function has been called, \code{\link[=bde_series_load_all]{bde_series_load_all()}} and
\code{\link[=bde_series_load]{bde_series_load()}} would try to find the series on the downloaded files.
}
\examples{
\dontshow{if (bde_check_access()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
bde_themes_download()
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=bde_series_load_all]{bde_series_load_all()}}

Other series: 
\code{\link{bde_series_full_load}()},
\code{\link{bde_series_load}()}
}
\concept{series}
