% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_scales.R
\name{bde_scale_colour_vivid}
\alias{bde_scale_colour_vivid}
\alias{bde_scale_color_vivid}
\alias{bde_scale_fill_vivid}
\alias{bde_scale_colour_rose}
\alias{bde_scale_color_rose}
\alias{bde_scale_fill_rose}
\title{Scales for \code{ggplot2}}
\usage{
bde_scale_colour_vivid(...)

bde_scale_color_vivid(...)

bde_scale_fill_vivid(...)

bde_scale_colour_rose(...)

bde_scale_color_rose(...)

bde_scale_fill_rose(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}. These scales are deprecated in favour of
\code{\link[=scale_color_bde_d]{scale_color_bde_d()}} and similar functions, in order to align the naming
with the current practices (see \code{\link[ggplot2:scale_viridis]{ggplot2::scale_fill_viridis_c()}}) and
similar functions).
}
\keyword{internal}
