% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-tidybde.R
\encoding{UTF-8}
\name{theme_tidybde}
\alias{theme_tidybde}
\title{BdE theme for \CRANpkg{ggplot2}}
\usage{
theme_tidybde(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[ggplot2:theme_classic]{ggplot2::theme_classic}}
  \describe{
    \item{\code{base_size}}{base font size, given in pts.}
    \item{\code{base_family}}{base font family}
    \item{\code{header_family}}{font family for titles and headers. The default, \code{NULL},
uses theme inheritance to set the font. This setting affects axis titles,
legend titles, the plot title and tag text.}
    \item{\code{base_line_size}}{base size for line elements}
    \item{\code{base_rect_size}}{base size for rect elements}
    \item{\code{ink,paper,accent}}{colour for foreground, background, and accented elements respectively.}
  }}
}
\value{
A \CRANpkg{ggplot2} \link[ggplot2:theme]{theme} object.
}
\description{
This \CRANpkg{ggplot2} theme is based on BdE publications.
}
\details{
This theme extends \code{\link[ggplot2:theme_classic]{ggplot2::theme_classic()}}.
}
\examples{
\dontshow{if (bde_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
library(ggplot2)
library(dplyr)
library(tidyr)

series_TC <- bde_series_full_load("TC_1_1.csv")

# Plot if the download succeeds.
if (nrow(series_TC) > 0) {
  series_TC <- series_TC[c(1, 2)]

  series_TC_pivot <- series_TC |>
    filter(
      Date >= "2020-01-01" & Date <= "2020-12-31",
      !is.na(series_TC[[2]])
    )

  names(series_TC_pivot) <- c("x", "y")

  ggplot(series_TC_pivot, aes(x = x, y = y)) +
    geom_line(linewidth = 0.8, color = bde_tidy_palettes(n = 1)) +
    labs(
      title = "Title",
      subtitle = "Some metric",
      caption = "Banco de España"
    ) +
    theme_tidybde()
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[ggplot2:theme_classic]{ggplot2::theme_classic()}} for the base theme extended here.

Plotting functions:
\code{\link[=bde_tidy_palettes]{bde_tidy_palettes()}},
\code{\link{scales_bde}}
}
\concept{bde_plot}
