% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalogs.R
\encoding{UTF-8}
\name{bde_catalog_update}
\alias{bde_catalog_update}
\title{Update BdE catalogs}
\source{
\href{https://www.bde.es/webbde/en/estadis/infoest/descarga_series_temporales.html}{Time-series bulk data download}
}
\usage{
bde_catalog_update(catalog = "ALL", cache_dir = NULL, verbose = FALSE)
}
\arguments{
\item{catalog}{A vector of characters indicating the catalogs to be updated
or \code{"ALL"} as a shorthand. See \strong{Details}.}

\item{cache_dir}{A path to a cache directory. The directory can also be set
via options with \code{options(bde_cache_dir = "path/to/dir")}.}

\item{verbose}{Logical \code{TRUE} or \code{FALSE}, display information useful for
debugging.}
}
\value{
None Downloads the catalog file(s) to the local machine.
}
\description{
Update the time-series catalogs provided by BdE.
}
\details{
Accepted values for \code{catalog} are:\tabular{llll}{
   CODE \tab PUBLICATION \tab UPDATE FREQUENCY \tab FREQUENCY \cr
   BE \tab Statistical Bulletin \tab Daily \tab Monthly \cr
   CF \tab Financial Accounts of the Spanish Economy \tab Quarterly \tab Annual \cr
   IE \tab Economic Indicators \tab Daily \tab Monthly \cr
   SI \tab Summary Indicators \tab Daily \tab Daily \cr
   TC \tab Exchange Rates \tab Daily \tab Daily \cr
   TI \tab Interest Rates \tab Daily \tab Daily \cr
   PB \tab Bank Lending Survey \tab Quarterly \tab Quarterly \cr
}


Use \code{"ALL"} as a shorthand for updating all the catalogs at a glance.
}
\examples{
\donttest{
bde_catalog_update("TI", verbose = TRUE)
}
}
\seealso{
Other catalog: 
\code{\link{bde_catalog_load}()}
}
\concept{catalog}
