% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bde_check_access.R
\encoding{UTF-8}
\name{bde_check_access}
\alias{bde_check_access}
\title{Check BdE access}
\usage{
bde_check_access()
}
\value{
A logical value indicating whether BdE resources are reachable.
}
\description{
Check whether \strong{R} can access resources at
\url{https://www.bde.es/webbe/en/estadisticas/recursos/descargas-completas.html}.
}
\examples{
\donttest{
bde_check_access()
}
}
\keyword{internal}
