% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bde_check_access.R
\name{bde_check_access}
\alias{bde_check_access}
\title{Check access to BdE}
\usage{
bde_check_access()
}
\value{
a logical.
}
\description{
Check if \strong{R} has access to resources at
\url{https://www.bde.es/webbe/en/estadisticas/recursos/descargas-completas.html}.
}
\examples{
\donttest{
bde_check_access()
}
}
\keyword{internal}
