% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bde_check_access}
\alias{bde_check_access}
\title{Check access to BdE}
\usage{
bde_check_access()
}
\value{
a logical.
}
\description{
Check if R has access to resources at
\url{https://www.bde.es/webbde/en/estadis/infoest/descarga_series_temporales.html}.
}
\examples{
\donttest{
bde_check_access()
}
}
\seealso{
Other utils: 
\code{\link{bde_parse_dates}()}
}
\concept{utils}
