% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superseded.R
\encoding{UTF-8}
\name{bde_vivid_pal}
\alias{bde_vivid_pal}
\alias{bde_rose_pal}
\title{Superseded BdE palettes}
\usage{
bde_vivid_pal(...)

bde_rose_pal(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
A color palette function.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These palettes have been superseded. Use \code{\link[=bde_tidy_palettes]{bde_tidy_palettes()}} instead.
}
\examples{
# Show the vivid palette.
scales::show_col(bde_vivid_pal()(6), labels = FALSE)

# Show the rose palette.
scales::show_col(bde_rose_pal()(6), labels = FALSE)
}
\keyword{internal}
