% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals_bde.R
\name{bde_vivid_pal}
\alias{bde_vivid_pal}
\alias{bde_scale_colour_vivid}
\alias{bde_scale_color_vivid}
\alias{bde_scale_fill_vivid}
\alias{bde_rose_pal}
\alias{bde_scale_colour_rose}
\alias{bde_scale_color_rose}
\alias{bde_scale_fill_rose}
\title{BdE colour palettes}
\usage{
bde_vivid_pal(...)

bde_scale_colour_vivid(...)

bde_scale_color_vivid(...)

bde_scale_fill_vivid(...)

bde_rose_pal(...)

bde_scale_colour_rose(...)

bde_scale_color_rose(...)

bde_scale_fill_rose(...)
}
\arguments{
\item{...}{Further arguments of the functions.}
}
\value{
A palette of colors.
}
\description{
Custom palettes based on the publications of BdE.
}
\examples{

# BdE vivid pal
scales::show_col(bde_vivid_pal()(6), labels = FALSE)

# BdE rose pal
scales::show_col(bde_rose_pal()(6), labels = FALSE)
}
\seealso{
\code{\link[=theme_bde]{theme_bde()}}
}
\concept{bde_plot}
