% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superseded.R
\name{bde_vivid_pal}
\alias{bde_vivid_pal}
\alias{bde_rose_pal}
\title{BdE superseded palettes}
\usage{
bde_vivid_pal(...)

bde_rose_pal(...)
}
\arguments{
\item{...}{Further arguments of the functions.}
}
\value{
A palette of colors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These functions have been superseded, see \code{\link[=bde_tidy_palettes]{bde_tidy_palettes()}} as a
replacement.

Custom palettes based on the publications of BdE.
}
\examples{

# BdE vivid pal
scales::show_col(bde_vivid_pal()(6), labels = FALSE)

# BdE rose pal
scales::show_col(bde_rose_pal()(6), labels = FALSE)
}
\keyword{internal}
