% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bde_vivid_pal}
\alias{bde_vivid_pal}
\alias{bde_rose_pal}
\title{BdE superseeded palettes}
\usage{
bde_vivid_pal(...)

bde_rose_pal(...)
}
\arguments{
\item{...}{Further arguments of the functions.}
}
\value{
A palette of colors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} See \code{\link[=bde_palettes]{bde_palettes()}}
}
\examples{

# BdE vivid pal
scales::show_col(bde_vivid_pal()(6), labels = FALSE)

# BdE rose pal
scales::show_col(bde_rose_pal()(6), labels = FALSE)
}
\keyword{internal}
