% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pals_bde.R
\name{bde_vivid_pal}
\alias{bde_vivid_pal}
\alias{bde_scale_colour_vivid}
\alias{bde_scale_color_vivid}
\alias{bde_scale_fill_vivid}
\title{BdE colour palettes}
\usage{
bde_vivid_pal(...)

bde_scale_colour_vivid(...)

bde_scale_color_vivid(...)

bde_scale_fill_vivid(...)
}
\arguments{
\item{...}{Further arguments of the functions.}
}
\description{
Custom \code{ggplot2} theme based on the publications of BdE.
}
\examples{

scales::show_col(bde_vivid_pal()(6))
}
\seealso{
\code{\link[=theme_bde]{theme_bde()}}
}
\concept{bde_plot}
