% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bde_pals.R
\name{bde_pals}
\alias{bde_pals}
\alias{bde_vivid_pal}
\alias{bde_rose_pal}
\title{BdE color palettes}
\usage{
bde_vivid_pal(...)

bde_rose_pal(...)
}
\arguments{
\item{...}{Further arguments of the functions.}
}
\value{
A palette of colors.
}
\description{
Custom palettes based on the publications of BdE.
}
\examples{

# BdE vivid pal
scales::show_col(bde_vivid_pal()(6), labels = FALSE)

# BdE rose pal
scales::show_col(bde_rose_pal()(6), labels = FALSE)
}
\seealso{
\code{\link[=theme_bde]{theme_bde()}}
}
\concept{bde_plot}
