% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_api.R
\encoding{UTF-8}
\name{bde_series_api}
\alias{bde_series_api}
\alias{bde_series_api_latest}
\title{Load BdE time series using the statistics web service (API)}
\usage{
bde_series_api_latest(series_alias, language = c("en", "es"), verbose = FALSE)
}
\arguments{
\item{series_alias}{Character string or vector of time series alias from the
\code{Nombre_de_la_serie} field of the corresponding series. See
\code{\link[=bde_catalog_load]{bde_catalog_load()}}.}

\item{language}{Character string. It can can take the values \code{"es"} or
\code{"en"}, to obtain the result in Spanish or English, respectively.}

\item{verbose}{Logical. If \code{TRUE}, display information useful for debugging.}
}
\value{
\code{bde_series_api_latest()} returns a \link[tibble:tbl_df]{tibble} with the latest
published data supplemented with other variables such as frequency,
description, or trend.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions are similar to \code{\link[=bde_series_load]{bde_series_load()}} but they use the \href{https://www.bde.es/webbe/en/estadisticas/recursos/api-estadisticas-bde.html}{Statistics web service (API)}.

\code{\link[=bde_series_api_latest]{bde_series_api_latest()}} allows you to obtain the latest published data for
one or more series.

This API service uses the series alias as identifier (see
\code{vignette("csv_manual.qmd", package = "tidyBdE")}). The series alias is a
positional code showing the location (column and/or row) of the series in
the table. However, although it is unique, it is not a good candidate to be
used as the series ID, as it is subject to change.
\itemize{
\item If a series changes position in the table, its alias will also change.
\item If a new series is inserted into a table, all the aliases that come after
it will change and will not represent the same series.
}
}
\examples{
\dontshow{if (bde_check_access()) withAutoprint(\{ # examplesIf}
\donttest{
xr <- bde_catalog_load(catalog = "TC")

# Extract latest value
library(dplyr)
xr |>
  slice_head(n = 3) |>
  pull(Nombre_de_la_serie) |>
  bde_series_api_latest(language = "en") |>
  glimpse()
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=bde_catalog_load]{bde_catalog_load()}},
\code{\link[=bde_catalog_search]{bde_catalog_search()}}, \code{\link[=bde_indicators]{bde_indicators()}}

Other series:
\code{\link[=bde_series_full_load]{bde_series_full_load()}},
\code{\link[=bde_series_load]{bde_series_load()}}
}
\concept{series}
