% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bde-tidy-palettes.R
\encoding{UTF-8}
\name{bde_tidy_palettes}
\alias{bde_tidy_palettes}
\title{BdE color palettes}
\usage{
bde_tidy_palettes(
  n = 6,
  palette = c("bde_vivid_pal", "bde_rose_pal", "bde_qual_pal"),
  alpha = NULL,
  rev = FALSE
)
}
\arguments{
\item{n}{The number of colors to return. Must be at least \code{1}.}

\item{palette}{A palette name: \code{"bde_vivid_pal"}, \code{"bde_rose_pal"} or
\code{"bde_qual_pal"}.}

\item{alpha}{Alpha transparency level in the range \verb{[0, 1]}, where \code{0} is
transparent and \code{1} is opaque. If \code{alpha = NULL}, the function does not
append opacity codes (\code{"FF"}) to individual color hex codes. See
\code{\link[ggplot2:alpha]{ggplot2::alpha()}}.}

\item{rev}{Logical. If \code{TRUE}, reverse the color order.}
}
\value{
A character vector of hex color codes.
}
\description{
These palettes are manually defined based on BdE publications. Each palette
contains at most six colors.
}
\examples{
# Show the BdE vivid palette.
scales::show_col(bde_tidy_palettes(palette = "bde_vivid_pal"),
  labels = FALSE
)

# Show the BdE rose palette.
scales::show_col(bde_tidy_palettes(palette = "bde_rose_pal"),
  labels = FALSE
)

# Show the BdE qualitative palette.
scales::show_col(bde_tidy_palettes(palette = "bde_qual_pal"),
  labels = FALSE
)
}
\seealso{
Plotting functions:
\code{\link{scales_bde}},
\code{\link[=theme_tidybde]{theme_tidybde()}}
}
\concept{bde_plot}
