% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bde_tidy_palettes.R
\encoding{UTF-8}
\name{bde_tidy_palettes}
\alias{bde_tidy_palettes}
\title{BdE color palettes}
\usage{
bde_tidy_palettes(
  n = 6,
  palette = c("bde_vivid_pal", "bde_rose_pal", "bde_qual_pal"),
  alpha = NULL,
  rev = FALSE
)
}
\arguments{
\item{n}{The number of colors (\verb{>= 1}) to return.}

\item{palette}{A valid palette name.}

\item{alpha}{An alpha transparency level in the range \verb{[0, 1]} (\code{0} means
transparent and \code{1} means opaque). If missing (i.e., \code{alpha = NULL}), the
function does not append opacity codes (\code{"FF"}) to the individual color
hex codes. See \code{\link[ggplot2:alpha]{ggplot2::alpha()}}.}

\item{rev}{Logical indicating whether to reverse the color order.}
}
\value{
A character vector of hex color codes.
}
\description{
Manually defined palettes based on BdE publications. Each palette contains
at most six colors.
}
\examples{

# Show the BdE vivid palette.
scales::show_col(bde_tidy_palettes(palette = "bde_vivid_pal"),
  labels = FALSE
)

# Show the BdE rose palette.
scales::show_col(bde_tidy_palettes(palette = "bde_rose_pal"),
  labels = FALSE
)

# Show the BdE qualitative palette.
scales::show_col(bde_tidy_palettes(palette = "bde_qual_pal"),
  labels = FALSE
)
}
\seealso{
Other bde_plot:
\code{\link{scales_bde}},
\code{\link[=theme_tidybde]{theme_tidybde()}}
}
\concept{bde_plot}
