% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_bde_d}
\alias{scale_color_bde_d}
\alias{scale_fill_bde_d}
\alias{scale_color_bde_c}
\alias{scale_fill_bde_c}
\title{BdE scales for \code{ggplot2}.}
\usage{
scale_color_bde_d(palette = "bde_vivid_pal", ...)

scale_fill_bde_d(palette = "bde_vivid_pal", ...)

scale_color_bde_c(palette = "bde_vivid_pal", ...)

scale_fill_bde_c(palette = "bde_vivid_pal", ...)
}
\arguments{
\item{palette}{Name of the BdE palette to apply. One of "bde_vivid_pal",
"bde_rose_pal". See \code{\link[=bde_vivid_pal]{bde_vivid_pal()}} for details.}

\item{...}{Further arguments of \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}}
}
\description{
Scales to be used with the \code{ggplot2} package. Discrete palettes are named
as \verb{scale_*_bde_d} while continuous palettes are named \verb{scale_*_bde_c}.
}
\seealso{
\code{\link[=bde_vivid_pal]{bde_vivid_pal()}}
}
\concept{bde_plot}
