% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBdE-package.R
\docType{package}
\name{tidyBdE-package}
\alias{tidyBdE}
\alias{tidyBdE-package}
\title{tidyBdE: Retrieve Time Series Data from 'Banco de España'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for retrieving 'Banco de España' ('BdE') time series data as 'tibble' objects from bulk CSV files and the 'Statistics web service (API)'. Bulk CSV functions use stable 'BdE' sequential numbers, while API functions use API series codes. Catalog functions support discovery and local caching. Plotting helpers provide 'ggplot2' palettes, scales and themes. 'Banco de España' is the national central bank and, within the framework of the Single Supervisory Mechanism ('SSM'), the supervisor of the Spanish banking system alongside the European Central Bank. This package is not sponsored, endorsed or administered by 'Banco de España'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenspain.github.io/tidyBdE/}
  \item \url{https://github.com/rOpenSpain/tidyBdE}
  \item Report bugs at \url{https://github.com/rOpenSpain/tidyBdE/issues}
}

}
\author{
\strong{Maintainer}: Diego H. Herrero \email{dev.dieghernan@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego H. Herrero \email{dev.dieghernan@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Banco de España (\href{https://ror.org/02f26yq04}{ROR}) (Data copyright holder) [copyright holder]
}

}
\keyword{internal}
