% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBdE-package.R
\docType{package}
\name{tidyBdE-package}
\alias{tidyBdE}
\alias{tidyBdE-package}
\title{tidyBdE: Retrieve Data from 'Banco de España'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for retrieving time series data from 'Banco de España' ('BdE') as 'tibble' objects. 'Banco de España' is the national central bank and, within the framework of the Single Supervisory Mechanism ('SSM'), the supervisor of the Spanish banking system alongside the European Central Bank. This package is not sponsored, endorsed or administered by 'Banco de España'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenspain.github.io/tidyBdE/}
  \item \url{https://github.com/rOpenSpain/tidyBdE}
  \item Report bugs at \url{https://github.com/rOpenSpain/tidyBdE/issues}
}

}
\author{
\strong{Maintainer}: Diego H. Herrero \email{dev.dieghernan@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Diego H. Herrero \email{dev.dieghernan@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID}) [copyright holder]
}

}
\keyword{internal}
