% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyBdE-package.R
\docType{package}
\name{tidyBdE-package}
\alias{tidyBdE}
\alias{tidyBdE-package}
\title{tidyBdE: Download Data from Bank of Spain}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Tools to download data series from 'Banco de España' ('BdE') on 
    'tibble' format. 'Banco de España' is the national central bank 
    and, within the framework of the Single Supervisory Mechanism ('SSM'), the 
    supervisor of the Spanish banking system along with the European Central 
    Bank. This package is in no way sponsored endorsed or administered by 
    'Banco de España'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dieghernan.github.io/tidyBdE/}
  \item \url{https://github.com/dieghernan/tidyBdE}
  \item Report bugs at \url{https://github.com/dieghernan/tidyBdE/issues}
}

}
\author{
\strong{Maintainer}: Diego H. Herrero \email{dev.dieghernan@gmail.com} (\href{https://orcid.org/0000-0001-8457-4658}{ORCID})

}
\keyword{internal}
